Address nit comments + move json to telemetry directory
